requirements.txt
streamlit>=1.34.0
openai>=1.51.2
pdfplumber>=0.11.0
python-docx>=1.1.2
